requirements.txt
# Core
python-dotenv==1.0.1
httpx==0.27.0
schedule==1.2.2

# Stock Data
yfinance==0.2.40
pandas>=2.0.0
numpy>=1.24.0
ta==0.11.0          # Technical Analysis (RSI, MACD, BB etc.)

# AI / LLM
openai==1.30.1

# LINE Messaging API
line-bot-sdk>=3.5.0

# Supabase
supabase==2.4.6

# News / Sentiment
feedparser==6.0.11
beautifulsoup4==4.12.3
requests>=2.31.0

# Scheduler / Cron
APScheduler==3.10.4

# Web Server (LINE Webhook受信)
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Machine Learning
scikit-learn>=1.4.0
joblib>=1.3.0
